meta-qti-ml: Install package config file of tensorflow
Issue: NNstreamer uses meson build system which for package config
uses pkgconfig.pc.in. If tensorflow-lite filter is enabled in
nnstreamer then it looks for tflite pkgconfig file and if not
available meson configuration is failed.

Fix: Add pkgconfig file for tensorflow-llite compilation, this will
ensure nnstreamer package config for tflite is successful.

Change-Id: I12f9eb16f430358b4e941e278f8313b0829c7711